requirements.txt
flask
schedule
RPi.GPIO  # Uncomment this line when deploying to Raspberry Pi
